Move example files to src.
I don't remember why I put these in FastImage in the first place.
