+++ ReactNativeFastImageExample/index.js
@@ -1,4 +1,4 @@
 import { AppRegistry } from 'react-native'
-import App from './FastImage/App'
+import App from './src/App'
 
 AppRegistry.registerComponent('ReactNativeFastImageExample', () => App)
