+++ ReactNativeFastImageExample/src/BorderRadiusExample.js
