+++ ReactNativeFastImageExample/src/Button.js
