+++ ReactNativeFastImageExample/src/FastImageExamples.js
