+++ ReactNativeFastImageExample/src/FeatureText.js
