+++ ReactNativeFastImageExample/src/GifExample.js
