+++ ReactNativeFastImageExample/src/Icons/Icon.js
