+++ ReactNativeFastImageExample/src/ProgressExample.js
