+++ ReactNativeFastImageExample/src/Section.js
