+++ ReactNativeFastImageExample/src/SectionFlex.js
