+++ ReactNativeFastImageExample/src/StatusBarUnderlay.js
